% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seam-carving.R
\name{calc_energy}
\alias{calc_energy}
\title{calc_energy}
\usage{
calc_energy(img)
}
\arguments{
\item{img}{EBImage Image class object}
}
\value{
EBImage Image
}
\description{
Function for calculating the energy map.
Energy is calculated by using horizontal and vertical sobel kernals.
}
