% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seam-carving.R
\name{checkUserInput}
\alias{checkUserInput}
\title{checkUserInput}
\usage{
checkUserInput(img, ncols, nrows)
}
\arguments{
\item{img}{EBImage Image class object}

\item{ncols}{Number by which the width has to be decreased}

\item{nrows}{Number by which the height has to be decreased}
}
\description{
Function for validating the user input that throws errors on incorrect values.
}
