% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/duplicate_seam_in_img.R
\name{duplicate_seam_in_img}
\alias{duplicate_seam_in_img}
\title{duplicate_seam_in_img}
\usage{
duplicate_seam_in_img(image, seam)
}
\arguments{
\item{image}{Image}

\item{seam}{array of pixel coordinates}
}
\value{
Image with one duplicated pixel per row
}
\description{
Function for duplicating a given seam in an image, causing the image to increase in size
}
