% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lowest_k_indices.R
\name{lowest_k_indices}
\alias{lowest_k_indices}
\title{lowest_k_indices}
\usage{
lowest_k_indices(vec, k, rev = FALSE)
}
\arguments{
\item{vec}{vector of last  entries of the cost map}

\item{k}{number of entries to return}

\item{rev}{boolean, true if image has been rotated}
}
\value{
vector with k lowest entries
}
\description{
Function returning the indices of the lowest k entries in the bottom row of the cost map.
}
