% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seam-carving.R
\name{mark_Seam}
\alias{mark_Seam}
\title{mark_seam}
\usage{
mark_Seam(img)
}
\arguments{
\item{img}{EBImage Image class object}
}
\value{
Image <EBImage class>
}
\description{
Function for replacing the values of the lowest cost seam in the image with NA's.
Image is reshaped to new dimensions after.
}
