% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seam-carving.R
\name{minimum_seam}
\alias{minimum_seam}
\title{minimum_seam}
\usage{
minimum_seam(img, ncol, nrow)
}
\arguments{
\item{img}{EBImage Image class object}

\item{ncol}{Number by which the width has to be decreased}

\item{nrow}{Number by which the height has to be decreased}
}
\value{
list(EnergyMap , BacktrackMap)
}
\description{
Function for calculating the minimal cost map and a backtrack map for finding seams.
}
