% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_image.R
\name{parseImage}
\alias{parseImage}
\title{parseImage}
\usage{
parseImage(imgPath)
}
\arguments{
\item{imgPath}{Path to the image}
}
\value{
Image <EBImage class>
}
\description{
Function for reading an image from a given directory path
}
