% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_seam.R
\name{remove_columns}
\alias{remove_columns}
\title{remove_columns}
\usage{
remove_columns(img, n)
}
\arguments{
\item{img}{EBImage Image class object}

\item{n}{Number by which the width has to be}
}
\value{
Image <EBImage class>
}
\description{
Function for removing n lowest cost seams in image. (Vertical)
}
