% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_seam.R
\name{remove_rows}
\alias{remove_rows}
\title{remove_rows}
\usage{
remove_rows(img, n)
}
\arguments{
\item{img}{EBImage Image class object}

\item{n}{Number by which the height has to be decreased}
}
\value{
Image <EBImage class>
}
\description{
Function for removing n lowest cost seams in image (Horizontal)
}
