% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove_seam.R
\name{remove_seam}
\alias{remove_seam}
\title{remove_seam}
\usage{
remove_seam(img)
}
\arguments{
\item{img}{EBImage Image}
}
\value{
Image <EBImage class>
}
\description{
Function for omitting all NA's in the image, which represent the seam.
Image is reshaped to new dimensions after.
}
