% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seam-carving-package.R
\docType{package}
\name{sc}
\alias{sc}
\alias{sc-package}
\title{Seam Carving - Content Aware Image Resizing}
\description{
The Seam Carving package provides functionalities for changing image dimensions by removing or adding 6-connected seams of pixels.
}
\section{Functions}{

sc_reduce(...)

sc_increase(...)

sc_mark_leastCSeam(...)
}

