% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_increase.R
\name{sc_increase}
\alias{sc_increase}
\title{sc_increase}
\usage{
sc_increase(imgPath, ncols = 0, nrows = 0)
}
\arguments{
\item{imgPath}{Path to a RGB image. (.PNG / .JPG / .TIF)}

\item{ncols}{Number by which the width has to be increased}

\item{nrows}{Number by which the height has to be increased}
}
\value{
SC-Image <Resized Image>
}
\description{
Increases the dimensions of an image by duplicating a given number of least cost seams
}
\examples{
library(EBImage)
filePath <- system.file("images", "sample-color.png", package="EBImage")
increased_image <- sc_increase(filePath, 10, 10)
display(increased_image$increased_img)
class(increased_image)
}
