% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_mark_leastCSeam.R
\name{sc_mark_leastCSeam}
\alias{sc_mark_leastCSeam}
\title{sc_mark_leastCSeam}
\usage{
sc_mark_leastCSeam(imgPath)
}
\arguments{
\item{imgPath}{Path to a three channel image. (.PNG / .JPG / .TIF)}
}
\value{
SC-SImage <EnergyMap & Image with marked seam>
}
\description{
Marks the least cost seam in vertical direction.
}
\examples{
library(EBImage)
filePath <- system.file("images", "sample-color.png", package="EBImage")
seam_image <- sc_mark_leastCSeam(filePath)
display(seam_image$energy_map)
display(seam_image$seam_image)
class(seam_image)
}
