% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seam-carving-package.R
\docType{package}
\name{SeamCarving}
\alias{SeamCarving}
\alias{SeamCarving-package}
\title{Seam Carving: A package for Content Aware Image Resizing}
\description{
The Seam Carving package provides a function for reducing image dimensions by removing seams and a function for
marking the least cost seam in vertical direction.
}
\section{Functions}{

sc_remove(...)

sc_leastCSeam(...)
}

