% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_energy.R
\name{calc_energy}
\alias{calc_energy}
\title{calc_energy}
\usage{
calc_energy(img)
}
\arguments{
\item{img}{EBImage Image class object}
}
\value{
Array
}
\description{
Function for calculating the energy map of a given EBImage Image.
Energy is calculated by using horizontal and vertical sobel kernals.
}
