% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_user_input.R
\name{checkUserInput}
\alias{checkUserInput}
\title{checkUserInput}
\usage{
checkUserInput(img, ncols, nrows)
}
\arguments{
\item{img}{EBImage Image object}

\item{ncols}{Number by which the width is reduced}

\item{nrows}{Number by which the height is reduced}
}
\description{
Function for validating the user input that throws errors on incorrect values.
}
