% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mark_seam.R
\name{mark_Seam}
\alias{mark_Seam}
\title{mark_seam}
\usage{
mark_Seam(img, mark = FALSE)
}
\arguments{
\item{img}{Image}

\item{mark}{Boolean, if TRUE the seam is not replaced with NA's but with ones.}
}
\value{
Image <EBImage class>
}
\description{
Function for replacing the values of the lowest cost seam in the image with NA's.
Image is reshaped to new dimensions after.
}
