% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/min_cost_map.R
\name{min_cost_map}
\alias{min_cost_map}
\title{min_cost_map}
\usage{
min_cost_map(img)
}
\arguments{
\item{img}{EBImage Image class object}
}
\value{
list(EnergyMap , BacktrackMap)
}
\description{
Function for calculating the minimal cost map and a backtrack map for finding seams.
}
