% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sc_reduce.R
\name{sc_reduce}
\alias{sc_reduce}
\title{Seam Carving <Reduce>}
\usage{
sc_reduce(imgPath, ncols = 0, nrows = 0)
}
\arguments{
\item{imgPath}{Path to a RGB image. (.PNG / .JPG / .TIF)}

\item{ncols}{Number by which the width has to be decreased}

\item{nrows}{Number by which the height has to be decreased}
}
\value{
SC-Image <EnergyMap & Resized Image>
}
\description{
Removes least energy vertical and/or horizontal pixel seams to reduce the size of a given image with respect to its contents.
}
\examples{
library(EBImage)
filePath <- system.file("images", "sample-color.png", package="EBImage")
reducedImage <- sc_reduce(filePath, ncols = 1, nrows = 1)
display(reducedImage$energy_map)
display(reducedImage$reduced_img)
class(reducedImage)
}
