% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seam-carving.R
\docType{package}
\name{SeamCarving}
\alias{SeamCarving}
\alias{SeamCarving-package}
\title{Seam Carving: A package for Content Aware Image Resizing}
\description{
For now, the Seam Carving package provides a function for reducing image dimensions by removing seams.
}
\section{Functions}{

sc_remove(...)
}

